deploy/ess_guardian_runtime/opt/opt/ess-guardian/current/thermal/requirements.txt
action-msgs==1.2.2
actionlib-msgs==4.9.0
Adafruit-Blinka==8.68.0
adafruit-circuitpython-busdevice==5.2.14
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-mlx90640==1.3.7
adafruit-circuitpython-register==1.11.1
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-typing==1.12.3
adafruit-extended-bus==1.0.2
Adafruit-PlatformDetect==3.85.0
Adafruit-PureIO==1.1.11
ament-clang-format==0.12.14
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-mypy==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-pyflakes==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
binho-host-adapter==0.1.6
bond==4.1.2
builtin-interfaces==1.2.2
cartographer-ros-msgs==2.0.9002
composition-interfaces==1.2.2
cv-bridge==3.2.1
diagnostic-msgs==4.9.0
diagnostic-updater==4.0.6
domain-coordinator==0.10.0
dwb-msgs==1.1.20
dynamixel-sdk==3.8.4
geometry-msgs==4.9.0
interactive-markers==2.3.2
laser-geometry==2.4.1
launch==1.0.13
launch-ros==0.19.13
launch-testing==1.0.13
launch-testing-ros==0.19.13
launch-xml==1.0.13
launch-yaml==1.0.13
lgpio==0.2.2.0
lifecycle-msgs==1.2.2
map-msgs==2.1.0
message-filters==4.3.11
my-hardware-pkg==0.0.0
nav-2d-msgs==1.1.20
nav-msgs==4.9.0
nav2-common==1.1.20
nav2-msgs==1.1.20
nav2-simple-commander==1.0.0
osrf-pycommon==2.1.6
paho-mqtt==2.1.0
pcl-msgs==1.0.0
pyftdi==0.57.1
pyserial==3.5
pyusb==1.3.1
rcl-interfaces==1.2.2
rclpy==3.3.19
rcutils==5.1.7
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.15
ros2bag==0.15.15
ros2cli==0.18.15
ros2component==0.18.15
ros2doctor==0.18.15
ros2interface==0.18.15
ros2launch==0.19.13
ros2lifecycle==0.18.15
ros2multicast==0.18.15
ros2node==0.18.15
ros2param==0.18.15
ros2pkg==0.18.15
ros2run==0.18.15
ros2service==0.18.15
ros2topic==0.18.15
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.6
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.3
rosidl-typesupport-fastrtps-cpp==2.2.3
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpi-lgpio==0.6
rpyutils==0.2.2
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
slam-toolbox==2.6.10
smbus2==0.5.0
smclib==4.1.2
sros2==0.10.7
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
sysv_ipc==1.1.0
tf2-geometry-msgs==0.25.17
tf2-kdl==0.25.17
tf2-msgs==0.25.17
tf2-py==0.25.17
tf2-ros-py==0.25.17
tf2-sensor-msgs==0.25.17
tf2-tools==0.25.17
theora-image-transport==2.5.4
toml==0.10.2
trajectory-msgs==4.9.0
turtlebot3-example==2.3.4
turtlebot3-msgs==2.4.0
turtlebot3-teleop==2.3.4
typing_extensions==4.15.0
unique-identifier-msgs==2.2.1
visualization-msgs==4.9.0
xacro==2.1.1
